requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
pbr>=0.6,!=0.7,<1.0

argparse
Babel>=1.3
python-heatclient>=0.3.0
oslo.config>=1.6.0  # Apache-2.0
oslo.i18n>=1.3.0  # Apache-2.0
pyOpenSSL>=0.11
six>=1.7.0
